Online services display basic file info

+++ src/vanillapdf_online_services/Controllers/ValidateController.cs
@@ -1,14 +1,16 @@
 ﻿using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
+using System;
 using System.IO;
+using System.Text;
 using vanillapdf.net;
 using vanillapdf_online_services.Common;
 using vanillapdf_online_services.Models.ViewModels;
 
 namespace vanillapdf_online_services.Controllers
 {
-    public class ValidateController : Controller
+    public class ValidateController : BaseController
     {
         private readonly ILogger _Logger;
         private readonly ICustomLocalizer _Loc;
@@ -45,6 +47,9 @@ namespace vanillapdf_online_services.Controllers
             string downloadPath = Path.Combine("TemporaryDownloads", UploadedFile.FileName);
 
             using (var uploadStream = UploadedFile.OpenReadStream()) {
+                string dirPath = Path.GetDirectoryName(downloadPath);
+                Directory.CreateDirectory(dirPath);
+
                 using (var fileStream = new FileStream(downloadPath, FileMode.Create)) {
                     uploadStream.CopyTo(fileStream);
 
@@ -56,16 +61,43 @@ namespace vanillapdf_online_services.Controllers
             // TODO: check db
 
             string hashPath = Path.Combine("FileDownloads", fileHash);
-            System.IO.File.Move(downloadPath, hashPath);
+            if (!System.IO.File.Exists(hashPath)) {
+                string dirPath = Path.GetDirectoryName(hashPath);
+                Directory.CreateDirectory(dirPath);
+
+                System.IO.File.Move(downloadPath, hashPath);
+            }
+
+            if (System.IO.File.Exists(downloadPath)) {
+                System.IO.File.Delete(downloadPath);
+            }
+
+            StringBuilder sb = new StringBuilder();
+            sb.AppendLine($"Filename: {UploadedFile.FileName}");
+            sb.AppendLine($"SHA-256: {fileHash}");
 
             using (var file = PdfFile.Open(hashPath)) {
                 file.Initialize();
 
                 using (var document = PdfDocument.OpenFile(file)) {
+                    var catalog = document.GetCatalog();
+
+                    var version = catalog.GetVersion();
+                    var pageTree = catalog.GetPageTree();
+
+                    int pageCount = pageTree.GetPageCount();
+
+                    string versionString = (version.HasValue ? Convert.ToString(version.Value) : "Not specified");
+
+                    sb.AppendLine($"Version: {versionString}");
+                    sb.AppendLine($"Pages: {pageCount}");
                 }
             }
 
-            return RedirectToAction(nameof(Index));
+            ViewData["ValidatedFile"] = sb.ToString();
+
+            SuccessMessage("File was successfully validated");
+            return View();
         }
     }
 }
